requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
email-validator==2.2.0
Werkzeug==3.1.3
pywebview==5.4
